Set proper rev for processed deleted doc
View changes needs to return the "winning" rev to the client. With
default style, it does this by looking up the rev in the seq tree. Prior
to this commit, when a document was being deleted, the default rev (ie.
{0, []}) would be added to the seq tree. This commit fixes the problem
by manually setting the proper rev in the deleted #doc record before
it's processed.